hemut-backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
python-jose[cryptography]
bcrypt
